chip_package.py: Get the pin geometry correct.
So far, all the devices have been square and so the fact that we had the
width and height dimensions reversed didn't matter.  We will shortly add
support for rectangular devices (TSSOP, SO8N - both of which are 2 x N
devices) and in these cases the geometry will matter when we do array
lookups.